image.gif,MKI.timex.py
So the gif is an optional feature, it's Rosey the Riveter giving you a nice arm pump when you complete your task. If you choose not to display it, you'll have to comment that labelframe out in the, "TimeX.task_complete", function. Will write a readme here shortly after I eat some breakfasts. This is a work in progress :] so if you feel like helping out, adding features, or other such sich, please shoot me a message or the like.